1. Split CappedIterable into an immutable and a mutable trait. 2. Renamed CappedSeq to CappedBuffer 3. Implemented CappedSet (both mutable and immutable) and CappedVector.
Still needs tests. More importantly, there is a conundrum of resolving the name clash between the immutable and mutable collections, at least the interfaces.
Moreover, I am not sure that a CappedSet should really extend a Set, CappedSeq a Seq, etc. On one hand, it increases flexibility, on another it breaks the contract regarding addition.